refactor: signout쪽 useAuthHandler에서 참조하는 것으로 변경
