+++ src/fsd_shared/hooks/index.ts
@@ -2,5 +2,4 @@ export { useInfiniteScroll } from './useInfiniteScroll';
 export { usePreviousValue } from './usePreviousValue';
 export { useIsWindowLg } from './useIsWindowLg';
 export { useTruncateParagraph } from './useTruncateParagraph';
-export { useHasAuth } from './useHasAuth';
-export { useErrorHandler } from './useErrorHandler';
+export { useAuthHandler } from './useAuthHandler';
