+++ src/widget/SideBar.tsx
@@ -10,27 +10,23 @@ import { NotificationWidget } from '@/fsd_widgets/notification';
 
 import { useUserInfo } from '@/fsd_entities/user';
 import { useMyInfoStore } from '@/fsd_entities/user';
-
-import { ProfileImage, SubHeader } from '@/entities';
+import { SubHeader } from '@/fsd_shared';
 import { Button } from '@/shadcn/components/ui';
-import { AuthRscService } from '@/shared';
+import { ProfileImage } from '@/fsd_shared/ui';
+import { useAuthHandler } from '@/fsd_shared/hooks';
 
 interface SideBarProps {
   className?: string;
 }
 
 export const SideBar = ({ className }: SideBarProps) => {
-  const { signout } = AuthRscService();
   const { updateMyInfoStore } = useUserInfo();
+  const { redirectToLogin } = useAuthHandler();
 
   const name = useMyInfoStore((state) => state.name);
   const email = useMyInfoStore((state) => state.email);
   const profileImage = useMyInfoStore((state) => state.profileImageUrl);
 
-  const handleNoRefresh = async () => {
-    await signout();
-    location.href = '/auth/signin';
-  };
 
   useEffect(() => {
     updateMyInfoStore();
@@ -42,7 +38,7 @@ export const SideBar = ({ className }: SideBarProps) => {
         size="icon"
         variant="ghost"
         className="absolute top-3 left-3 flex cursor-pointer flex-col gap-2 p-0 text-black shadow-none xl:top-4 xl:right-4 xl:left-auto"
-        onClick={handleNoRefresh}
+        onClick={redirectToLogin}
       >
         <LogOut className="size-6 xl:size-8" />
         <p className="hidden text-xs font-light xl:block xl:text-sm">로그아웃</p>
